updating interfaces and cancel-surprese route

+++ src/app/api/cancel-surprise/route.ts
@@ -7,13 +7,22 @@ export async function POST(request: Request) {
   const { surpriseId } = await request.json()
 
   try {
+    // Check if the surprise exists
+    const existingSurprise = await prisma.surprise.findUnique({
+      where: { id: surpriseId },
+    })
+
+    if (!existingSurprise) {
+      return NextResponse.json({ error: 'Surpresa não encontrada' }, { status: 404 })
+    }
+
     await prisma.$transaction(async (prisma) => {
-      // Deletar todos os tickets associados à surpresa
+      // Delete all tickets associated with the surprise
       await prisma.tickets.deleteMany({
         where: { idSurprise: surpriseId },
       })
 
-      // Deletar a surpresa
+      // Delete the surprise
       await prisma.surprise.delete({
         where: { id: surpriseId },
       })
@@ -23,5 +32,7 @@ export async function POST(request: Request) {
   } catch (error) {
     console.error('Erro ao deletar surpresa e tickets:', error)
     return NextResponse.json({ error: 'Erro ao deletar surpresa e tickets' }, { status: 500 })
+  } finally {
+    await prisma.$disconnect()
   }
 }
